rename stuff, hardware diagnostics => diagnostics
