+++ apps/admin/frontend/src/screens/diagnostics_screen.test.tsx
